chore(release): migrate plugin and cli tag tracks
