+++ .gitattributes
@@ -1,0 +1,1 @@
+*.sh text eol=lf
