+++ .github/workflows/ci-dashboard.yml
@@ -5,12 +5,14 @@ on:
     branches: [main, dev]
     paths:
       - "plugin/dashboard/**"
+      - "scripts/check-plugin-version-sync.py"
       - "scripts/check-server-version-sync.py"
       - ".github/workflows/ci-dashboard.yml"
   pull_request:
     branches: [main, dev]
     paths:
       - "plugin/dashboard/**"
+      - "scripts/check-plugin-version-sync.py"
       - "scripts/check-server-version-sync.py"
       - ".github/workflows/ci-dashboard.yml"
 
@@ -49,8 +51,8 @@ jobs:
         with:
           python-version: "3.11"
 
-      - name: Verify server-owned version metadata
-        run: python scripts/check-server-version-sync.py
+      - name: Verify plugin-owned version metadata
+        run: python scripts/check-plugin-version-sync.py
 
       - name: Install dashboard API test deps
         run: pip install -r relay_server/requirements.txt fastapi httpx pytest requests
