+++ desktop/src/updater.ts
@@ -3,7 +3,8 @@
 //
 // Strategy:
 //   1. Query the GitHub Releases API for this repo (default Codename-11/hermes-relay),
-//      filter to tags starting with `desktop-v`, take the newest. Zero deps.
+//      prefer tags starting with `cli-v`, and fall back to historical
+//      `desktop-v` prereleases. Zero deps.
 //   2. Pick the asset for this platform (`process.platform` + `process.arch`).
 //   3. Download to `<target>.download`, verify against SHA256SUMS.txt, then
 //      atomically rename — POSIX rename keeps the running process's inode
@@ -84,9 +85,10 @@ interface ParsedVersion {
 }
 
 function parseVersion(raw: string): ParsedVersion | null {
-  // Strip leading `v` / `desktop-v`.
+  // Strip leading `v` / `cli-v` / historical `desktop-v`.
   let v = raw.trim()
-  if (v.startsWith('desktop-v')) v = v.slice('desktop-v'.length)
+  if (v.startsWith('cli-v')) v = v.slice('cli-v'.length)
+  else if (v.startsWith('desktop-v')) v = v.slice('desktop-v'.length)
   else if (v.startsWith('v')) v = v.slice(1)
 
   const [core, pre] = v.split('-', 2)
@@ -101,6 +103,13 @@ function parseVersion(raw: string): ParsedVersion | null {
   }
   const prerelease = pre ? pre.split('.').filter(Boolean) : []
   return { major, minor, patch, prerelease }
+}
+
+function releaseVersionFromTag(tag: string): string {
+  if (tag.startsWith('cli-v')) return tag.slice('cli-v'.length)
+  if (tag.startsWith('desktop-v')) return tag.slice('desktop-v'.length)
+  if (tag.startsWith('v')) return tag.slice(1)
+  return tag
 }
 
 /** Returns 1 if a > b, -1 if a < b, 0 if equal. SemVer §11 precedence rules:
@@ -187,19 +196,22 @@ export async function checkForUpdate(opts: { repo?: string } = {}): Promise<Upda
   const repo = opts.repo ?? DEFAULT_REPO
   const releases = await fetchReleases(repo)
 
-  // Filter to desktop-v* tags. NOTE: GitHub orders the response by the release
-  // row's created_at, NOT by SemVer of the tag — and "created_at" can shift
-  // when the row is touched (re-tag, manual edit). Don't trust [0]; pick the
-  // SemVer-maximum tag explicitly so a touched alpha.9 row can't outrank a
-  // freshly-tagged alpha.10. Caught by Bailey on 2026-04-25 — `update --check`
-  // reported "up to date" on alpha.9 because the API listed alpha.9 first.
-  const desktop = releases.filter((r) => r.tag_name.startsWith('desktop-v'))
-  if (desktop.length === 0) return null
-  const pick = desktop.reduce((max, r) =>
+  // Prefer cli-v* tags. Historical public prereleases used desktop-v*, so keep
+  // a fallback while the alpha channel migrates. NOTE: GitHub orders the
+  // response by the release row's created_at, NOT by SemVer of the tag — and
+  // "created_at" can shift when the row is touched (re-tag, manual edit).
+  // Don't trust [0]; pick the SemVer-maximum tag explicitly so a touched alpha
+  // row can't outrank a freshly-tagged release.
+  const cli = releases.filter((r) => r.tag_name.startsWith('cli-v'))
+  const candidates = cli.length > 0
+    ? cli
+    : releases.filter((r) => r.tag_name.startsWith('desktop-v'))
+  if (candidates.length === 0) return null
+  const pick = candidates.reduce((max, r) =>
     compareVersions(r.tag_name, max.tag_name) > 0 ? r : max
   )
 
-  const latestVersion = pick.tag_name.slice('desktop-v'.length)
+  const latestVersion = releaseVersionFromTag(pick.tag_name)
   const current = VERSION
 
   const wantAsset = assetNameForPlatform()
