+++ plugin/relay/__init__.py
@@ -13,12 +13,12 @@ See ``plugin/relay/server.py`` for the aiohttp server,
 # during its own import via ``from . import __version__``, so this ordering
 # avoids a circular-import crash during package initialization.
 #
-# Canonical server version source is pyproject.toml's [project].version.
-# Keep this runtime constant in sync with pyproject.toml for server-v*
+# Canonical plugin version source is pyproject.toml's [project].version.
+# Keep this runtime constant in sync with pyproject.toml for plugin-v*
 # releases. Android releases use gradle/libs.versions.toml and android-v* tags;
-# desktop CLI releases use desktop/package.json and desktop-v* tags. The
-# /health endpoint reports this server version, and stale values make live
-# server diagnosis harder than it should be.
+# CLI releases use desktop/package.json and cli-v* tags. The /health endpoint
+# reports this plugin version, and stale values make live diagnosis harder than
+# it should be.
 __version__ = "1.1.0"
 
 from .server import create_app, main  # noqa: E402 — must come after __version__
