+++ scripts/bump-relay-version.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
-# Backward-compatible wrapper for the renamed server release helper.
+# Backward-compatible wrapper for the renamed plugin release helper.
 
 set -euo pipefail
 
 REPO_ROOT="$(cd "$(dirname "$0")/.." && pwd)"
-echo "NOTE: scripts/bump-relay-version.sh is deprecated; use scripts/bump-server-version.sh."
-exec bash "$REPO_ROOT/scripts/bump-server-version.sh" "$@"
+echo "NOTE: scripts/bump-relay-version.sh is deprecated; use scripts/bump-plugin-version.sh."
+exec bash "$REPO_ROOT/scripts/bump-plugin-version.sh" "$@"
