+++ scripts/check-relay-version-sync.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-"""Backward-compatible wrapper for scripts/check-server-version-sync.py."""
+"""Backward-compatible wrapper for scripts/check-plugin-version-sync.py."""
 
 from __future__ import annotations
 
@@ -9,13 +9,13 @@ import sys
 
 
 REPO_ROOT = pathlib.Path(__file__).resolve().parents[1]
-TARGET = REPO_ROOT / "scripts" / "check-server-version-sync.py"
+TARGET = REPO_ROOT / "scripts" / "check-plugin-version-sync.py"
 
 
 if __name__ == "__main__":
     print(
         "NOTE: scripts/check-relay-version-sync.py is deprecated; "
-        "use scripts/check-server-version-sync.py.",
+        "use scripts/check-plugin-version-sync.py.",
         file=sys.stderr,
     )
     runpy.run_path(str(TARGET), run_name="__main__")
