+++ user-docs/desktop/index.md
@@ -11,7 +11,7 @@ It also includes a terminal escape hatch for when *you* want to drive: bare `her
 :::
 
 ::: info Where this track is headed
-This surface is focusing into a **remote-hands connector** — remote control, filesystem, and terminal access for the agent on machines you install it to. Desktop chat and management UX belong to [hermes-desktop](https://github.com/NousResearch/hermes-agent); this CLI's `chat` mode keeps working for scripting but isn't where new features land. "Desktop" is shorthand, not a constraint — the same binary runs on laptops and headless servers (`daemon` mode needs no display at all), and the track naming will be reframed to match in a later refactor.
+This surface is focusing into a **remote-hands connector** — remote control, filesystem, and terminal access for the agent on machines you install it to. Desktop chat and management UX belong to [hermes-desktop](https://github.com/NousResearch/hermes-agent); this CLI's `chat` mode keeps working for scripting but isn't where new features land. "Desktop" is shorthand, not a constraint — the same binary runs on laptops and headless servers (`daemon` mode needs no display at all). New release tags use the `cli-v*` track; historical alpha prereleases used `desktop-v*`.
 :::
 
 ## The point — the agent works on *your* machine
@@ -129,4 +129,4 @@ Use `shell` when you want to drive interactively; use `chat --json` from scripts
 - [Hermes Agent](https://github.com/NousResearch/hermes-agent) — the agent platform the CLI talks to.
 - [Herm](https://github.com/liftaris/herm) — OpenTUI dashboard plugin installable from the desktop surface.
 - [CLI GitHub source](https://github.com/Codename-11/hermes-relay/tree/main/desktop) — `@hermes-relay/cli` package.
-- [Release notes](https://github.com/Codename-11/hermes-relay/releases?q=desktop) — tagged `desktop-v*` (separate track from Android).
+- [Release notes](https://github.com/Codename-11/hermes-relay/releases?q=cli) — tagged `cli-v*` (separate track from Android); old alpha prereleases are under `desktop-v*`.
